Fix currency type defs
