+++ assets/js/types/type-defs/index.ts
@@ -2,3 +2,4 @@ export * from './cart-response';
 export * from './product-response';
 export * from './cart';
 export * from './hooks';
+export * from './currency';
