+++ bin/webpack-helpers.js
@@ -88,6 +88,7 @@ const getAlias = ( options = {} ) => {
 			__dirname,
 			`../assets/js/${ pathPart }previews/`
 		),
+		'@woocommerce/types': path.resolve( __dirname, `../assets/js/types/` ),
 	};
 };
