+++ packages/prices/index.js
@@ -1,2 +1,1 @@
 export * from './utils';
-export * from './types';
