--- packages/prices/types.ts
@@ -1,9 +1,0 @@
-export interface Currency {
-	code: string;
-	decimalSeparator: string;
-	minorUnit: number;
-	prefix: string;
-	suffix: string;
-	symbol: string;
-	thousandSeparator: string;
-}
