+++ tests/js/jest.config.json
@@ -27,7 +27,8 @@
 		"@woocommerce/resource-previews": "assets/js/previews",
 		"@woocommerce/shared-context": "assets/js/shared/context",
 		"@woocommerce/shared-hocs": "assets/js/shared/hocs",
-		"@woocommerce/blocks-test-utils": "tests/utils"
+		"@woocommerce/blocks-test-utils": "tests/utils",
+		"@woocommerce/types": "assets/js/types"
 	},
 	"setupFiles": [
 		"@wordpress/jest-preset-default/scripts/setup-globals.js",
